% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbugs-package.R
\docType{package}
\encoding{UTF-8}
\name{pbugs-package}
\alias{pbugs-package}
\alias{_PACKAGE}
\title{Run WinBUGS Chains in Parallel}
\description{
Call WinBUGS from your R session running multiple chains
  in parallel. Results are returned in an object of class \code{bugs} and
  \code{pbugs}.
}
\details{
Mantainer: Miguel Ángel Martínez-Beneito \email{martinez_mig@gva.es}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.org/fisabio/pbugs}
  \item Report bugs at \url{https://github.org/fisabio/pbugs/issues}
}

}
\author{
\strong{Maintainer}: Miguel Ángel Martínez-Beneito \email{martinez_mig@gva.es}

Authors:
\itemize{
  \item Gonzalo García-Donato \email{garcia_gon@gva.es}
  \item Paloma Botella-Rocamora \email{botella_pal@gva.es}
  \item Jordi Pérez-Panadés \email{perez_jorpan@gva.es}
  \item Paqui Corpas-Burgos \email{corpas_fra@gva.es}
  \item Carlos Vergara-Hernández \email{vergara_car@gva.es}
}

}
\keyword{package}
