% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.pbugs}
\alias{print.pbugs}
\title{Print \code{pbugs} Objects}
\usage{
\method{print}{pbugs}(x, digits.summary = 2, order.results = c("none",
  "Rhat", "n.eff"), ...)
}
\arguments{
\item{x}{Object of class \code{pbugs}.}

\item{digits.summary}{Number of digits to print.}

\item{order.results}{Should summary results be ordered by \code{Rhat} or
\code{n.eff}? No by default.}

\item{...}{Further arguments to be used for the \code{print} method for
\code{bugs} objects (\code{R2WinBUGS} library).}
}
\description{
S3 Method to print \code{pbugs} objects.
}
