% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbugs-package.R
\docType{data}
\name{sample_df}
\alias{sample_df}
\title{Sample data for a logistic regression}
\format{A \code{data.frame} with 1000 observations and 4 variables:
  \describe{
    \item{x1}{Binomial with prob = 0.7.}
    \item{x2}{Binomial with prob = 0.4.}
    \item{x3}{Normal with mean 40 and sd 10 (mean centered).}
    \item{y}{Response.}
  }}
\usage{
sample_df
}
\description{
The data creation script is in /data-raw directory.
}
\examples{
data(sample_df)
}
\keyword{datasets}
