% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbugs-package.R
\docType{package}
\encoding{UTF-8}
\name{pbugs-package}
\alias{pbugs-package}
\alias{_PACKAGE}
\title{Runs WinBUGS or OpenBUGS Models in Parallel}
\description{
Call WinBUGS or OpenBUGS from your R session running multiple
  chains in parallel. Results are returned in an object of class \code{bugs}
  and \code{pbugs}.
}
\details{
Mantainer: Carlos Vergara-Hernández \email{vergara_car@gva.es}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fisabio/pbugs}
  \item Report bugs at \url{https://github.com/fisabio/pbugs/issues}
}

}
\author{
\strong{Maintainer}: Carlos Vergara-Hernández \email{vergara_car@gva.es} (\href{https://orcid.org/0000-0002-7263-856X}{ORCID})

Authors:
\itemize{
  \item Miguel Ángel Martínez-Beneito \email{miguel.a.martinez@uv.es} (\href{https://orcid.org/0000-0001-8406-8050}{ORCID})
}

Other contributors:
\itemize{
  \item Andrew Gelman \email{gelman@stat.columbia.edu} (R2WinBUGS package author) [copyright holder]
  \item Sibylle Sturtz \email{sturtz@statistik.tu-dortmund.de} (R2WinBUGS package author) [copyright holder]
  \item Uwe Ligges \email{ligges@statistik.tu-dortmund.de} (R2WinBUGS package author) [copyright holder]
  \item Gregor Gorjanc \email{gregor.gorjanc@bfro.uni-lj.si} (R2WinBUGS package author) [copyright holder]
  \item Jouni Kerman \email{kerman@stat.columbia.edu} (R2WinBUGS package author) [copyright holder]
  \item Neal Thomas \email{snthomas99@gmail.com} (R2OpenBUGS package author) [copyright holder]
}

}
\keyword{package}
