% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbugs-package.R
\docType{package}
\encoding{UTF-8}
\name{pbugs-package}
\alias{pbugs-package}
\alias{_PACKAGE}
\title{Runs WinBUGS or OpenBUGS Models in Parallel}
\description{
Call WinBUGS or OpenBUGS from your R session running multiple
  chains in parallel. Results are returned in an object of class \code{bugs}
  and \code{pbugs}.
}
\details{
Mantainer: Carlos Vergara-Hernández \email{vergara_car@gva.es}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fisabio/pbugs}
  \item Report bugs at \url{https://github.com/fisabio/pbugs/issues}
}

}
\author{
\strong{Maintainer}: Carlos Vergara-Hernández \email{vergara_car@gva.es}

Authors:
\itemize{
  \item Miguel Ángel Martínez-Beneito \email{martinez_mig@gva.es}
}

}
\keyword{package}
