% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.pbugs}
\alias{print.pbugs}
\title{Print pbugs objects}
\usage{
\method{print}{pbugs}(x, digits.summary = 2, order_results = c("none",
  "Rhat", "n.eff"), ...)
}
\arguments{
\item{x}{Object of class \code{pbugs}.}

\item{digits.summary}{Number of digits to print.}

\item{order_results}{Should summary results be ordered by Rhat or n.eff?}

\item{...}{Further arguments to use in \code{print}.}
}
\description{
S3 Method to print pbugs objects.
}
