% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{traceplot}
\alias{traceplot}
\title{Trace plot of MCMC draws}
\usage{
pbugs_trace(x, cutoff_neff = 100, cutoff_rhat = 1.1, ...)
}
\arguments{
\item{x}{Object of class \code{bugs} or \code{pbugs}.}

\item{cutoff_neff}{Maximum threshold of effective sample size to select parameters.}

\item{cutoff_rhat}{Minimum threshold of Rhat to select parameters.}

\item{...}{Additional arguments to be passed to \code{\link[bayesplot]{mcmc_trace}} function.}
}
\description{
Trace plot of MCMC draws, following Neff & Rhat criteria
}
\seealso{
\code{\link[bayesplot]{mcmc_trace}}
}
