% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{traceplot}
\alias{traceplot}
\title{Trace plot of MCMC draws}
\usage{
traceplot(x, parameter_name = character(), parameter_pattern = character(),
only_not_converged = FALSE,  max_pars = 9, ...)
}
\arguments{
\item{x}{Object of class \code{bugs} or \code{pbugs}.}

\item{parameter_name}{Literal name of the parameters to select.}

\item{parameter_pattern}{Regular expression (as used in \code{\link[base]{grep}}) for parameter selection.}

\item{only_not_converged}{Select only parameters that did not converged? Dafault = FALSE.}

\item{max_pars}{Maximum number of parameter to show. Default: 9.}

\item{...}{Additional arguments to be passed to \code{\link[graphics]{plot}} function.}
}
\description{
Trace plot of MCMC draws that did not converge following Neff & Rhat criteria
}
