% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{traceplot}
\alias{traceplot}
\title{Trace plot of MCMC draws}
\usage{
traceplot(x, cutoff_neff = 100, cutoff_rhat = 1.1, max_pars = 9, mfrow = c(3, 3), ...)
}
\arguments{
\item{x}{Object of class \code{bugs} or \code{pbugs}.}

\item{cutoff_neff}{Maximum threshold of effective sample size to select parameters.}

\item{cutoff_rhat}{Minimum threshold of Rhat to select parameters.}

\item{max_pars}{Maximum number of parameter to show. Default: 9.}

\item{mfrow}{Number of rows and columns in the plot.}

\item{...}{Additional arguments to be passed to \code{\link[coda]{traceplot}} function.}
}
\description{
Trace plot of MCMC draws that did not converge following Neff & Rhat criteria
}
